refactor, add foodItem
