+++ src/main/java/com/robertx22/age_of_exile/database/data/loot_chest/base/LootChestData.java
@@ -8,13 +8,14 @@ import com.robertx22.age_of_exile.gui.texts.ExileTooltips;
 import com.robertx22.age_of_exile.gui.texts.textblocks.AdditionalBlock;
 import com.robertx22.age_of_exile.gui.texts.textblocks.NameBlock;
 import com.robertx22.age_of_exile.gui.texts.textblocks.RarityBlock;
-import com.robertx22.age_of_exile.gui.texts.textblocks.LeveledItemLevelBlock;
+import com.robertx22.age_of_exile.gui.texts.textblocks.usableitemblocks.UsageBlock;
 import com.robertx22.age_of_exile.saveclasses.gearitem.gear_bases.TooltipContext;
 import com.robertx22.age_of_exile.uncommon.datasaving.StackSaving;
 import com.robertx22.age_of_exile.uncommon.interfaces.data_items.ICommonDataItem;
 import com.robertx22.age_of_exile.uncommon.localization.Chats;
 import com.robertx22.age_of_exile.uncommon.localization.ChestContent;
 import com.robertx22.age_of_exile.uncommon.localization.Itemtips;
+import com.robertx22.age_of_exile.uncommon.utilityclasses.TooltipUtils;
 import com.robertx22.library_of_exile.utils.ItemstackDataSaver;
 import com.robertx22.library_of_exile.vanilla_util.main.VanillaUTIL;
 import net.minecraft.ChatFormatting;
@@ -96,19 +97,19 @@ public class LootChestData implements ICommonDataItem<GearRarity> {
         tip.addAll(new ExileTooltips()
                 .accept(new NameBlock(Collections.singletonList(ctx.stack.getHoverName())))
                 .accept(new RarityBlock(getRarity()))
-                .accept(new AdditionalBlock(ImmutableList.of(
-                        Itemtips.CHEST_CONTAINS.locName(new ChestContent(getLootChest().GUID()).get().locName()))
+                .accept(new UsageBlock(ImmutableList.of(
+                        Itemtips.CHEST_CONTAINS.locName(new ChestContent(getLootChest().GUID()).get().locName().withStyle(ChatFormatting.YELLOW)),
+                        TooltipUtils.level(lvl))
                 ))
-                .accept(new AdditionalBlock(ImmutableList.of(Chats.OPEN_LOOT_CHEST.locName().withStyle(ChatFormatting.GRAY))))
+                .accept(new AdditionalBlock(ImmutableList.of(Chats.OPEN_LOOT_CHEST.locName().withStyle(ChatFormatting.AQUA))))
                 .accept(new AdditionalBlock(() -> {
                     //weird
                     if (getKeyItem() != null) {
-                        return ImmutableList.of(Itemtips.NEED_KEY.locName(getKeyItem().getDefaultInstance().getHoverName()));
+                        return ImmutableList.of(Itemtips.NEED_KEY.locName(getKeyItem().getDefaultInstance().getHoverName()).withStyle(ChatFormatting.GOLD));
                     } else {
                         return EMPTY_LIST;
                     }
                 }).showWhen(this::isLocked))
-                .accept(new LeveledItemLevelBlock(lvl))
                 .release());
