+++ src/main/java/com/robertx22/age_of_exile/gui/texts/ExileTooltips.java
@@ -1,7 +1,6 @@
 package com.robertx22.age_of_exile.gui.texts;
 
 import com.robertx22.age_of_exile.gui.texts.textblocks.*;
-import com.robertx22.age_of_exile.gui.texts.textblocks.usableitemblocks.DragableBlock;
 import com.robertx22.age_of_exile.uncommon.utilityclasses.TooltipUtils;
 import com.robertx22.library_of_exile.wrappers.ExileText;
 import lombok.NoArgsConstructor;
@@ -58,7 +57,7 @@ public class ExileTooltips {
                         collect.get(BlockCategories.REQUIREMENT),
                         collect.get(BlockCategories.STAT),
                         collect.get(BlockCategories.USAGE),
-                        collect.get(BlockCategories.LEVELED_ITEM_LEVEL)
+                        collect.get(BlockCategories.LEVELED_ITEM)
                         )
                 .filter(Objects::nonNull)
                 .flatMap(Collection::stream)
@@ -71,7 +70,6 @@ public class ExileTooltips {
         //handle additional blocks, the order of aBs is the putting order.
         List<AbstractTextBlock> additions = collect.get(BlockCategories.ADDITIONAL);
         if (additions != null){
-            additions.sort(Comparator.comparing(x -> x instanceof DragableBlock));
             for (AbstractTextBlock abstractTextBlock : additions) {
                 if (abstractTextBlock != null && !abstractTextBlock.getAvailableComponents().isEmpty()) {
                     list.addAll(abstractTextBlock.getAvailableComponents());
@@ -86,11 +84,9 @@ public class ExileTooltips {
                         collect.get(BlockCategories.DURABILITY)
                 )
                 .filter(Objects::nonNull)
-                .flatMap(Collection::stream)
+                .map(x -> x.get(0))
                 .filter(x -> !x.getAvailableComponents().isEmpty())
-                .forEachOrdered(x -> {
-                    list.addAll(x.getAvailableComponents());
-                });
+                .forEachOrdered(x -> list.addAll(x.getAvailableComponents()));
 
         list.add(emptyLine);
         Optional.ofNullable(collect.get(BlockCategories.OPERATION))
@@ -117,7 +113,7 @@ public class ExileTooltips {
         STAT,
         DURABILITY,
         USAGE,
-        LEVELED_ITEM_LEVEL,
+        LEVELED_ITEM,
         OPERATION,
         ADDITIONAL
