+++ src/main/java/com/robertx22/age_of_exile/saveclasses/jewel/JewelItemData.java
@@ -6,7 +6,10 @@ import com.robertx22.age_of_exile.database.data.rarities.GearRarity;
 import com.robertx22.age_of_exile.database.registry.ExileDB;
 import com.robertx22.age_of_exile.gui.inv_gui.actions.auto_salvage.ToggleAutoSalvageRarity;
 import com.robertx22.age_of_exile.gui.texts.ExileTooltips;
-import com.robertx22.age_of_exile.gui.texts.textblocks.*;
+import com.robertx22.age_of_exile.gui.texts.textblocks.NameBlock;
+import com.robertx22.age_of_exile.gui.texts.textblocks.OperationTipBlock;
+import com.robertx22.age_of_exile.gui.texts.textblocks.RarityBlock;
+import com.robertx22.age_of_exile.gui.texts.textblocks.RequirementBlock;
 import com.robertx22.age_of_exile.gui.texts.textblocks.affixdatablocks.SimpleItemStatBlock;
 import com.robertx22.age_of_exile.mmorpg.registers.common.items.RarityItems;
 import com.robertx22.age_of_exile.mmorpg.registers.common.items.SlashItems;
@@ -107,13 +110,16 @@ public class JewelItemData implements ICommonDataItem<GearRarity>, IStatCtx {
                 .accept(new RarityBlock(getRarity()))
                 .accept(new SimpleItemStatBlock(info)
 
-                        .acceptIf(Itemtips.JEWEL_STATS.locName().withStyle(ChatFormatting.BLUE), affixes.stream().flatMap(x -> x.getAllStatsWithCtx(lvl, info).stream()).toList(), this.auraStats.isEmpty())
+                        .acceptIf(Itemtips.JEWEL_STATS.locName().withStyle(ChatFormatting.BLUE),
+                                affixes.stream().flatMap(x -> x.getAllStatsWithCtx(lvl, info).stream()).toList(),
+                                this.auraStats.isEmpty())
 
-                        .acceptIf(Itemtips.UNIQUE_STATS.locName().withStyle(ChatFormatting.YELLOW), this.auraStats.stream().flatMap(x -> x.getStatAndContext(ClientOnly.getPlayer()).stream().flatMap(statContext -> statContext.stats.stream())).toList(), !this.auraStats.isEmpty())
+                        .acceptIf(Itemtips.UNIQUE_STATS.locName().withStyle(ChatFormatting.YELLOW),
+                                this.auraStats.stream().flatMap(x -> x.getStatAndContext(ClientOnly.getPlayer()).stream().flatMap(statContext -> statContext.stats.stream())).toList(),
+                                !this.auraStats.isEmpty())
 
                         .accept(Itemtips.COR_STATS.locName().withStyle(ChatFormatting.RED), cor.stream().flatMap(x -> x.getAllStatsWithCtx(lvl, info).stream()).toList())
                 )
-                .accept(new RarityBlock(this.getRarity()))
                 .accept(new RequirementBlock(this.lvl))
                 .accept(new OperationTipBlock().setShift().setAlt())
                 .release());
