+++ src/main/java/com/robertx22/age_of_exile/saveclasses/stat_soul/StatSoulData.java
@@ -11,7 +11,6 @@ import com.robertx22.age_of_exile.gui.texts.ExileTooltips;
 import com.robertx22.age_of_exile.gui.texts.textblocks.AdditionalBlock;
 import com.robertx22.age_of_exile.gui.texts.textblocks.OperationTipBlock;
 import com.robertx22.age_of_exile.gui.texts.textblocks.RarityBlock;
-import com.robertx22.age_of_exile.gui.texts.textblocks.usableitemblocks.DragableBlock;
 import com.robertx22.age_of_exile.gui.texts.textblocks.usableitemblocks.UsageBlock;
 import com.robertx22.age_of_exile.loot.LootInfo;
 import com.robertx22.age_of_exile.loot.blueprints.GearBlueprint;
@@ -241,7 +240,8 @@ public class StatSoulData implements ICommonDataItem<GearRarity>, ISettableLevel
 
         ExileTooltips exileTooltips = new ExileTooltips()
                 .accept(new RarityBlock(getRarity()))
-                .accept(new DragableBlock(DragableBlock.AvailableTarget.GEAR_SOUL));
+                .accept(new UsageBlock(Collections.singletonList(Chats.INFUSES_STATS.locName().withStyle(ChatFormatting.AQUA))));
+
         List<Component> tooltip = new ArrayList<>();
         if (this.gear != null) {
 
@@ -264,13 +264,11 @@ public class StatSoulData implements ICommonDataItem<GearRarity>, ISettableLevel
                             .locName().withStyle(ChatFormatting.BLUE)).withStyle(ChatFormatting.GRAY));
                 }
             }
+            exileTooltips.accept(new UsageBlock(tooltip));
         }
-        exileTooltips.accept(new UsageBlock(Collections.singletonList(Chats.INFUSES_STATS.locName().withStyle(ChatFormatting.AQUA))));
-        exileTooltips.accept(new UsageBlock(tooltip));
-
 
         exileTooltips.accept(new AdditionalBlock(Collections.singletonList(Chats.RIGHT_CLICK_TO_GEN_ITEM.locName().withStyle(ChatFormatting.BLUE))).showWhen(() -> cangen));
-        exileTooltips.accept(new OperationTipBlock().setAlt());
+        exileTooltips.accept(new OperationTipBlock().setAlt().addDraggableTipAbove(OperationTipBlock.AvailableTarget.GEAR));
         return exileTooltips;
     }
