+++ src/main/java/com/robertx22/age_of_exile/uncommon/localization/Itemtips.java
@@ -9,6 +9,7 @@ public enum Itemtips implements IAutoLocName {
     //add "tip" at the end to avoid same variable call
     //you can use \n in here, just remember use the splitLongText() to process it.
     //a special method splitLongText() in TooltipsUtils.java can split the sentence into different line by replacing the "\n", achieving a actual line break function.
+    BUFF_TIP("Buff Stats: "),
     SOUL_MODIFIER_TIP("Soul modifier can be applied to gear soul, which then generate only the corresponding gear type when use."),
     SOUL_MODIFIER_USE_TIP("[Drag onto gear soul to use]"),
     PROF_TOOL_LEVEL_CAP("Capped Drops to LVL %1$s"),
